requirements.txt
Pillow~=8.2.0
opencv-python~=4.5.1.48
mediapipe~=0.8.3.1
numpy~=1.20.2
pandas~=1.2.4